requirements.txt
fitz==0.0.1.dev2
keras==2.14.0
opencv_contrib_python==4.8.1.78
opencv_python==4.8.0.76
Pillow==9.4.0
Pillow==10.3.0
pydantic==1.10.15
ttkbootstrap==1.10.1
